requirements.txt
python-telegram-bot==13.12
matplotlib==3.4.3
pytz==2021.3
redis==4.3.3
requests==2.26.0
urllib3==1.26.7